isolate IMainProcessService and ISharedProcessService into common
related to #114418
