+++ src/vs/platform/ipc/electron-browser/mainProcessService.ts
@@ -5,7 +5,7 @@
 
 import { IChannel, IServerChannel, StaticRouter } from 'vs/base/parts/ipc/common/ipc';
 import { Server as MessagePortServer } from 'vs/base/parts/ipc/electron-browser/ipc.mp';
-import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/mainProcessService';
+import { IMainProcessService } from 'vs/platform/ipc/common/services';
 
 /**
  * An implementation of `IMainProcessService` that leverages MessagePorts.
