+++ src/vs/workbench/contrib/webview/electron-browser/webviewIgnoreMenuShortcutsManager.ts
@@ -9,7 +9,7 @@ import { DisposableStore, IDisposable, toDisposable } from 'vs/base/common/lifec
 import { isMacintosh } from 'vs/base/common/platform';
 import { ProxyChannel } from 'vs/base/parts/ipc/common/ipc';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { IMainProcessService } from 'vs/platform/ipc/electron-sandbox/mainProcessService';
+import { IMainProcessService } from 'vs/platform/ipc/common/services';
 import { IWebviewManagerService } from 'vs/platform/webview/common/webviewManagerService';
 import { WebviewMessageChannels } from 'vs/workbench/contrib/webview/browser/baseWebviewElement';
